requirements.txt
Flask
nltk
langdetect
requests
python-dotenv
chardet
scikit-learn
